+++ packages/metastable/src/comfy/session/models.ts
@@ -12,6 +12,37 @@ export class ComfyConditioning {
   ) {}
 }
 
+export class ComfyGuider {
+  constructor(
+    private session: ComfySession,
+    public ref: RPCRef,
+  ) {}
+
+  async sample(
+    {
+      noise,
+      sampler,
+      sigmas,
+      latent,
+    }: {
+      noise: RPCRef;
+      sampler: RPCRef;
+      sigmas: RPCRef;
+      latent: ComfyLatent;
+    },
+    preview?: ComfyPreviewSettings,
+  ) {
+    return (await this.session.invoke('sampling:sample_custom', {
+      noise,
+      guider: this.ref,
+      sampler,
+      sigmas,
+      latent,
+      preview,
+    })) as RPCRef;
+  }
+}
+
 export class ComfyCheckpoint {
   constructor(
     private session: ComfySession,
@@ -51,7 +82,7 @@ export class ComfyCheckpoint {
     },
     preview?: ComfyPreviewSettings,
   ) {
-    return (await this.session.invoke('checkpoint:sample', {
+    return (await this.session.invoke('sampling:sample', {
       diffusion_model: this.data.diffusionModel,
       latent,
       positive: conditioning.positive,
@@ -75,12 +106,46 @@ export class ComfyCheckpoint {
     })) as RPCRef[];
   }
 
+  async decodeTiled(
+    samples: RPCRef,
+    tileSize: number,
+    overlap: number,
+    temporalSize: number,
+    temporalOverlap: number,
+  ) {
+    return (await this.session.invoke('vae:decode_tiled', {
+      vae: this.data.vae,
+      samples,
+      tile_size: tileSize,
+      overlap,
+      temporal_size: temporalSize,
+      temporal_overlap: temporalOverlap,
+    })) as RPCRef[];
+  }
+
   async encode(image: RPCRef, mask?: RPCRef) {
     return (await this.session.invoke('vae:encode', {
       vae: this.data.vae,
       image,
       mask,
     })) as ComfyLatent;
+  }
+
+  async getGuider(conditioning: RPCRef) {
+    const ref = (await this.session.invoke('sampling:basic_guider', {
+      diffusion_model: this.data.diffusionModel,
+      conditioning,
+    })) as RPCRef;
+    return new ComfyGuider(this.session, ref);
+  }
+
+  async getSigmas(schedulerName: string, steps: number, denoise: number) {
+    return (await this.session.invoke('sampling:get_sigmas', {
+      diffusion_model: this.data.diffusionModel,
+      scheduler_name: schedulerName,
+      steps,
+      denoise,
+    })) as RPCRef;
   }
 }